se modifico  el css de la tabla
